Added a clone list
Previous version presented an unexpected behavior. It might be because the unduplicating_list is being iterated through and modified at the same time. A clone of the list is added to make sure that the unduplicating_list is only being iterated through, while the clone list is the one being modified