Added the ability to search for the number with the most duplicate
Previous version threw an error identifying the most frequently occurring input. This might be because the variable that stores the most frequently occurring input is not a list. This means that if two or more inputs have the same highest_value, the variable gets overwritten resulting in unexpected behaviours.

Another reason as to why it failed might be because the program run like this:
    search dictionary for keys with highest value -> if the current key has the highest value -> set it as the most reoccurring input (number with most duplicates)
> This presents a problem since I am setting the highest value while using it as a standard at the same time
 So the highest_value keeps changing as the program loops

which I renewed into this:
    search the dictionary for the highest value first -> then run another for loop and search for the key with the same corresponding value -> store all those keys in a list.
> This ensures that the highest_value remains constant