Added the ability to accept input and store them into lists and dicts
It accepts input from the user and stores it in a list. The list is then copied to a dictionary that assigns the input as the key, and the number of instances of that input as the value. If the number of instances exceeds 1, then the input is a duplicate