Revised the entire thing
Instead of using a split, modify, join approach, it instead used a
similar method as replicate_ lstrip.py. That is, by using a variable
called state that detects wether the previous string is a whitespace or
a non-whitespace.

If the current character is a whitespace, it will keep the state to zero
letting it know that it is a whitespace. If the previous character is a
whitespace, it means the current character that is being iterated is the
first letter in a word. This way, it knows what part of the string have
to be capitalized.