Removed problematic index() function
Added another way of tracking the index of the current iteration in a loop.
Now uses the len() function and a 'counter' variable to determine that.
Apparently, the index() function returns the index of the first
occurrence of an item instead of the actual index of the item. So if the
item has a replicate that came before, it returns that instead.