Replaced the nested for loop with something better
Instead of using a for loop...
First it identifies the length of the prefix to determine how much loops
will be made. It then uses an "index_mark" variable to move through the
characters in the text and prefix. As long as the characters match, keep
going. Once it fails to match, return False and exit the loop. However,
if the loop finishes and all of the characters in the prefix matches that
of the text, return True.